Improve performance of CLI generate.
Doing a `findAll()` and then counting the length of the array is highly enefficient.
Using `findAll()` slowing down the generation of large data sets for testing.
Change the code to an SQL count via the JPA `count()` method.

+++ src/main/java/org/tdl/vireo/cli/Cli.java
@@ -127,7 +127,7 @@ public class Cli implements CommandLineRunner {
 
                         {
                             Random random = new Random();
-                            int idOffset = cliService.countUsers();
+                            long idOffset = cliService.countUsers();
                             User helpfulHarry = null;
                             boolean hasSubmissionTypes = false;
 
@@ -136,7 +136,7 @@ public class Cli implements CommandLineRunner {
                                 hasSubmissionTypes = cliService.hasSubmissionTypes();
                             }
 
-                            for (int i = itemsGenerated; i < num1 + itemsGenerated; i++) {
+                            for (long i = itemsGenerated; i < num1 + itemsGenerated; i++) {
                                 cliService.operateGenerate(expansive, num2, random, idOffset, helpfulHarry, hasSubmissionTypes, i);
                                 System.out.print("\r" + (i - itemsGenerated) + " of " + num1 + " generated...");
                             }
